Move read_all method to read_all.rb ReadAll module
